requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.1.2
zipp==3.8.0
python-dotenv
requests>=2.20.0
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
urllib3==1.26.5
